+++ frontend/src/components/organisms/Homeranking.tsx
@@ -1,17 +1,17 @@
 import { FC } from 'react';
 import { Header } from 'semantic-ui-react';
 import styled from 'styled-components';
-import RankThreeCards from 'components/molecules/RankThreeCards';
+import HomeRankThreeCards from 'container/EnhancedHomeRankThreeCards';
 
 const Rankinggrid = styled.div`
   margin: 3rem 0 !important;
 `;
 
-const Homeranking: FC = () => (
+const HomeRanking: FC = () => (
   <Rankinggrid className="ui vertical stripe center aligned segment grid">
     <Header as="h2" content="お気に入りアイテムランキング" />
-    <RankThreeCards />
+    <HomeRankThreeCards />
   </Rankinggrid>
 );
 
-export default Homeranking;
+export default HomeRanking;
