+++ frontend/src/components/templates/Hometop.tsx
@@ -1,7 +1,7 @@
 import { FC } from 'react';
 import Hometoptitles from 'components/organisms/Hometoptitles';
 import Homeicongrid from 'components/organisms/Homeicongrid';
-import Homeranking from 'components/organisms/Homeranking';
+import Homeranking from 'components/organisms/HomeRanking';
 
 const Hometop: FC = () => (
   <>
