+++ frontend/src/hooks/HomeRankiThreeCards.tsx
